requirements.txt
argparse==1.2.1
distribute==0.6.24
django-countries==3.0.1
wsgiref==0.1.2
